requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.12
jinja2==3.1.4

pdfplumber==0.11.4
python-docx==1.1.2

groq>=0.4.0
python-dotenv>=1.0.0
spacy>=3.7.0

